Extracted the shared 'contextualize()' dispatch into a trait.
