+++ src/Stacks/AbstractStack.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace DrevOps\EnvironmentDetector\Stacks;
 
-use DrevOps\EnvironmentDetector\Contexts\ContextInterface;
-use DrevOps\EnvironmentDetector\Contexts\Drupal;
-use DrevOps\EnvironmentDetector\Contexts\DrupalContextualizerInterface;
+use DrevOps\EnvironmentDetector\DispatchesContextualization;
 
 /**
  * Abstract stack.
@@ -16,6 +14,8 @@ use DrevOps\EnvironmentDetector\Contexts\DrupalContextualizerInterface;
  * @package DrevOps\EnvironmentDetector\Stacks
  */
 abstract class AbstractStack implements StackInterface {
+
+  use DispatchesContextualization;
 
   /**
    * Stack ID. Stacks should override this constant.
@@ -27,28 +27,6 @@ abstract class AbstractStack implements StackInterface {
    */
   public function id(): string {
     return static::ID;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function contextualize(ContextInterface $context): void {
-    // Known interfaces - optimised for speed.
-    if ($this instanceof DrupalContextualizerInterface && $context instanceof Drupal) {
-      // Only the most-derived contextualizeDrupal() runs, so a subclass that
-      // overrides it must call parent::contextualizeDrupal() to keep the
-      // parent's settings.
-      $this->contextualizeDrupal($context);
-      return;
-    }
-
-    // Runtime.
-    $method = 'contextualize' . (new \ReflectionClass($context))->getShortName();
-    $callable = [$this, $method];
-
-    if (is_callable($callable)) {
-      $callable($context);
-    }
   }
 
 }
