+++ tests/Stacks/AbstractStackTest.php
@@ -6,12 +6,15 @@ namespace DrevOps\EnvironmentDetector\Tests\Stacks;
 
 use DrevOps\EnvironmentDetector\Contexts\Drupal;
 use DrevOps\EnvironmentDetector\Contexts\DrupalContextualizerInterface;
+use DrevOps\EnvironmentDetector\DispatchesContextualization;
 use DrevOps\EnvironmentDetector\Stacks\AbstractStack;
 use DrevOps\EnvironmentDetector\Tests\EnvironmentDetectorTestCase;
 use DrevOps\EnvironmentDetector\Tests\Fixtures\NotDrupalContext;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\CoversTrait;
 
 #[CoversClass(AbstractStack::class)]
+#[CoversTrait(DispatchesContextualization::class)]
 final class AbstractStackTest extends EnvironmentDetectorTestCase {
 
   public function testIdDefaultsToConst(): void {
